Organized Data directory into Models and Repositories Added Services directory Created empty IBracketGeneratorService and derived BracketGeneratorService Created empty INameGeneratorService and derived NameGeneratorService Added singleton instances of new services into DI services
