+++ Data/Models/Bracket.cs
@@ -1,0 +1,6 @@
+﻿namespace DndFightClubGeneratorSite.Data.Models;
+
+public class Bracket
+{
+    
+}
