+++ DndFightClubGeneratorSite.csproj
@@ -6,4 +6,8 @@
     <ImplicitUsings>enable</ImplicitUsings>
   </PropertyGroup>
 
+  <ItemGroup>
+    <Folder Include="Data\Repositories" />
+  </ItemGroup>
+
 </Project>
