+++ Pages/FetchData.razor
@@ -3,6 +3,8 @@
 <PageTitle>Weather forecast</PageTitle>
 
 @using DndFightClubGeneratorSite.Data
+@using DndFightClubGeneratorSite.Data.Models
+@using DndFightClubGeneratorSite.Services
 @inject WeatherForecastService ForecastService
 
 <h1>Weather forecast</h1>
