+++ Services/BracketGeneratiorService.cs
@@ -1,0 +1,5 @@
+﻿namespace DndFightClubGeneratorSite.Services;
+
+class BracketGenerationService : IBracketGenerationService
+{
+}
