+++ Services/INameGeneratorService.cs
@@ -1,0 +1,6 @@
+﻿namespace DndFightClubGeneratorSite.Services;
+
+public interface INameGeneratorService
+{
+    
+}
